+++ examples/all-polkavm-networks/hardhat.config.ts
@@ -6,15 +6,12 @@ const PRIVATE_KEY = process.env.PRIVATE_KEY
 
 const config: HardhatUserConfig = {
     solidity: "0.8.28",
-    resolc: {
-        compilerSource: "npm",
-    },
     networks: {
         // npx hardhat node
         hardhat: {
             polkavm: true,
             nodeConfig: {
-                nodeBinaryPath: "./bin/substrate-node",
+                nodeBinaryPath: "./bin/dev-node",
                 rpcPort: 8000,
                 dev: true,
             },
@@ -39,7 +36,7 @@ const config: HardhatUserConfig = {
         polkadotHubTestnet: {
             polkavm: true,
             url: "https://testnet-passet-hub-eth-rpc.polkadot.io",
-            accounts: !!PRIVATE_KEY
+            accounts: PRIVATE_KEY
                 ? [PRIVATE_KEY]
                 : ["271ad9a5e1e0178acebdb572f8755aac3463d863ddfc70e32e7d5eb0b334e687"],
             chainId: 420420422,
@@ -53,7 +50,8 @@ const config: HardhatUserConfig = {
         kusamaHub: {
             polkavm: true,
             url: "https://kusama-asset-hub-eth-rpc.polkadot.io",
-            accounts: !!PRIVATE_KEY ? [PRIVATE_KEY] : [],
+            polkadotUrl: "wss://kusama-asset-hub-rpc.polkadot.io",
+            accounts: PRIVATE_KEY ? [PRIVATE_KEY] : [],
             chainId: 420420418,
         },
         /** Internal Parity Testnet
@@ -65,7 +63,7 @@ const config: HardhatUserConfig = {
         westendHub: {
             polkavm: true,
             url: "https://westend-asset-hub-eth-rpc.polkadot.io",
-            accounts: !!PRIVATE_KEY
+            accounts: PRIVATE_KEY
                 ? [PRIVATE_KEY]
                 : ["271ad9a5e1e0178acebdb572f8755aac3463d863ddfc70e32e7d5eb0b334e687"],
             chainId: 420420421,