+++ packages/hardhat-polkadot-node/src/type-extensions.ts
@@ -47,11 +47,13 @@ declare module "hardhat/types/config" {
     interface HardhatNetworkConfig {
         polkavm?: boolean
         url?: string
+        polkadotUrl?: string
     }
 
     interface HttpNetworkConfig {
         polkavm?: boolean
         ethNetwork?: string
+        polkadotUrl?: string
     }
 }