+++ packages/hardhat-polkadot-node/src/types.ts
@@ -1,5 +1,6 @@
 import { HardhatNetworkForkingUserConfig } from "hardhat/types"
 import { HardhatNetworkUserConfig } from "hardhat/types/config"
+import { ChopsticksService, EthRpcService, SubstrateNodeService } from "./services"
 
 export interface CliCommands {
     nodeBinaryPath?: string
@@ -17,9 +18,15 @@ export interface CommandArguments {
     forkBlockNumber?: string | number
     nodeCommands?: HardhatNetworkUserConfig["nodeConfig"]
     adapterCommands?: HardhatNetworkUserConfig["adapterConfig"]
+    docker?: HardhatNetworkUserConfig["docker"]
 }
 
 export interface RpcServer {
+    services(): {
+        substrateNodeService: SubstrateNodeService | null
+        ethRpcService: EthRpcService | null
+        chopsticksService: ChopsticksService | null
+    }
     listen(chopsticksArgs?: string[], adapterArgs?: string[], blockProcess?: boolean): Promise<void>
     stop(): Promise<void>
 }